feat: implement MS-DOS mode and refactor command prompt
- Extracted command prompt logic into `DOSShell` class for reusability.
- Implemented MS-DOS mode triggered by `#msdos` URL hash.
- Enabled "Restart in MS-DOS mode" in the Shutdown dialog.
- Added full-screen DOS app launching support in MS-DOS mode.
- Blocked Windows apps in MS-DOS mode with accurate error message.
- Added `win` command to return to Windows GUI.
